+++ build_test.go
@@ -338,7 +338,7 @@ func Test_dockerBuilder_goBuildArgs(t *testing.T) {
 			},
 		},
 		{
-			name: "verbosity disabled, windows",
+			name: "verbosity disabled, windows/amd64",
 			fields: fields{
 				verbose: false,
 				pkg:     "fyne-io/fyne-example",
@@ -356,6 +356,29 @@ func Test_dockerBuilder_goBuildArgs(t *testing.T) {
 				"go", "build",
 				"-ldflags", "'-H windowsgui -X main.version=1.0.0'",
 				"-o", "build/test-windows-amd64.exe",
+				"-a",
+				"fyne-io/fyne-example",
+			},
+		},
+		{
+			name: "verbosity disabled, windows/386",
+			fields: fields{
+				verbose: false,
+				pkg:     "fyne-io/fyne-example",
+				workDir: "/code/test",
+				output:  "test",
+				ldflags: "-X main.version=1.0.0",
+			},
+			args: args{
+				target: "windows/386",
+			},
+			want: []string{
+				"-e", "CGO_ENABLED=1",
+				"-e", "GOOS=windows", "-e", "GOARCH=386", "-e", "CC=i686-w64-mingw32-gcc",
+				dockerImage,
+				"go", "build",
+				"-ldflags", "'-H windowsgui -X main.version=1.0.0'",
+				"-o", "build/test-windows-386.exe",
 				"-a",
 				"fyne-io/fyne-example",
 			},