+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Add wildcard support for goarch [#15](https://github.com/lucor/fyne-cross/issues/15)
+- Fix misleading error message when docker is not available [#19](https://github.com/lucor/fyne-cross/issues/19)
 
 ## [1.2.1] - 2019-10-26
 - Fix fyne-cross docker image build tag